Optimize memory management in ULT.
- limit the amount of atomic reads
- change debug callback code to not test atomic
- change logic of seeing if we are within limits.

+++ unit_tests/helpers/memory_management.h
@@ -64,7 +64,7 @@ struct AllocationEvent {
     void *callstack[CallStackSize];
     int fastLeakDetectionMode = 0;
 };
-enum : size_t {
+enum : int {
     maxEvents = 1024 * 1024,
     fastEvents = 1024 * 1024
 };
